refactor: 패키지 구조 변경
- constants 패키지를 common 패키지 내부로 변경
- utils 패키지를 common 패키지 내부로 변경
