+++ src/main/java/lotto/common/utils/Parser.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.common.utils;
 
 import java.util.Arrays;
 import java.util.List;
