+++ src/main/java/lotto/model/Money.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import static lotto.constants.LottoRules.PURCHASE_AMOUNT_UNIT;
+import static lotto.common.constants.LottoRules.PURCHASE_AMOUNT_UNIT;
 
 public class Money {
     private static final String AMOUNT_UNIT_EXCEPTION = "구입 금액은 " + PURCHASE_AMOUNT_UNIT + "원 단위로 입력해야합니다.";
