+++ src/test/java/lotto/common/utils/InputValidatorTest.java
@@ -1,8 +1,8 @@
-package lotto.utils;
+package lotto.common.utils;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import lotto.utils.InputValidator;
+import lotto.common.utils.InputValidator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
